+++ DevTools/Inspector/RemoteObjectGraphModel.cpp
@@ -64,7 +64,7 @@ GUI::ModelIndex RemoteObjectGraphModel::parent_index(const GUI::ModelIndex& inde
 
     // NOTE: If the parent has no parent, it's a root, so we have to look among the remote roots.
     if (!remote_object.parent->parent) {
-        for (int row = 0; row < m_process.roots().size(); ++row) {
+        for (size_t row = 0; row < m_process.roots().size(); ++row) {
             if (&m_process.roots()[row] == remote_object.parent)
                 return create_index(row, 0, remote_object.parent);
         }
@@ -72,7 +72,7 @@ GUI::ModelIndex RemoteObjectGraphModel::parent_index(const GUI::ModelIndex& inde
         return {};
     }
 
-    for (int row = 0; row < remote_object.parent->parent->children.size(); ++row) {
+    for (size_t row = 0; row < remote_object.parent->parent->children.size(); ++row) {
         if (&remote_object.parent->parent->children[row] == remote_object.parent)
             return create_index(row, 0, remote_object.parent);
     }
