+++ Libraries/LibCore/Object.cpp
@@ -102,7 +102,7 @@ void Object::insert_child_before(Object& new_child, Object& before_child)
 
 void Object::remove_child(Object& object)
 {
-    for (int i = 0; i < m_children.size(); ++i) {
+    for (size_t i = 0; i < m_children.size(); ++i) {
         if (m_children.ptr_at(i).ptr() == &object) {
             // NOTE: We protect the child so it survives the handling of ChildRemoved.
             NonnullRefPtr<Object> protector = object;
